Added a way to batch compare

+++ lib/filesrebuilder/GUI/ComparePointer.rb
@@ -7,30 +7,19 @@ module FilesRebuilder
       # Set the pointer to be compared in this widget
       #
       # Parameters::
-      # * *pointer* (_FileInfo_ or _SegmentPointer_): Pointer to be compared
-      # * *matching_info* (_MatchingInfo_): Matching information of all files matching this pointer
+      # * *itr_pointer* (_PointerIterator_): Pointer iterator giving pointers to be compared
       # * *matching_selection* (_MatchingSelection_): Current user selection of matching files
-      def set_pointer_to_compare(pointer, matching_info, matching_selection)
+      # * *single* (_Boolean_): Is there a single pointer to compare?
+      def set_pointers_to_compare(itr_pointer, matching_selection, single)
         @matching_selection = matching_selection
-        @idx_focused = nil
-        @builder['original_container'] << @gui_controller.create_widget_for_matching_pointer(pointer, matching_selection.matching_pointers[pointer] == pointer)
-        # For each encountered CRC, keep the matching pointer widget
-        # map< String, Gtk::Widget >
-        crcs = {}
-        matching_pointers_container = @builder['matching_container']
-        matching_info.matching_files(@gui_controller.options[:score_min]).sort_by { |_, matching_file_info| matching_file_info.score }.reverse_each do |matching_pointer, matching_file_info|
-          crc = matching_pointer.get_crc
-          if crcs.has_key?(crc)
-            # Add it to the existing widget
-            crcs[crc].add_pointer(matching_pointer)
-          else
-            matching_pointer_widget = @gui_controller.create_widget_for_matching_pointer(matching_pointer, matching_selection.matching_pointers[pointer] == matching_pointer)
-            matching_pointers_container << matching_pointer_widget
-            crcs[crc] = matching_pointer_widget
-          end
-        end
-        # First, focus the original one
-        set_focused(-1)
+        @itr_pointer = itr_pointer
+        @single = single
+        @builder['comparison_navigation_hbox'].visible = !single
+        goto_next
+      end
+
+      def on_close_button_clicked(button_widget)
+        self.destroy
       end
 
       def on_next_match_button_clicked(button_widget)
@@ -54,6 +43,14 @@ module FilesRebuilder
         select_pointer_widget(((idx_focused == -1) ? @builder['original_container'].children[0] : @builder['matching_container'].children[idx_focused]))
       end
 
+      def on_next_comparison_button_clicked(button_widget)
+        goto_next
+      end
+
+      def on_previous_comparison_button_clicked(button_widget)
+        goto_prev
+      end
+
       # Select a given Matching pointer widget
       #
       # Parameters::
@@ -62,10 +59,86 @@ module FilesRebuilder
         original_pointer = @builder['original_container'].children[0].matching_pointers[0]
         matched_pointer = matching_pointer_widget.matching_pointers[0]
         @matching_selection.matching_pointers[original_pointer] = matched_pointer
-        self.destroy
+        goto_next
       end
 
       private
+
+      # Display the next comparison
+      def goto_next
+        pointer, matching_pointers = @itr_pointer.next(@builder['skip_matched_checkbutton'].active?, true)
+        if (pointer == nil)
+          if @single
+            self.destroy
+          else
+            @itr_pointer.reset
+            pointer, matching_pointers = @itr_pointer.next(@builder['skip_matched_checkbutton'].active?, true)
+            if (pointer == nil)
+              notify('No more items to compare')
+            else
+              notify('Cycling items to compare from the beginning')
+            end
+          end
+        else
+          notify('')
+        end
+        if (pointer != nil)
+          load_comparison(pointer, matching_pointers)
+        end
+      end
+
+      # Display the previous comparison
+      def goto_prev
+        pointer, matching_pointers = @itr_pointer.next(@builder['skip_matched_checkbutton'].active?, false)
+        if (pointer == nil)
+          @itr_pointer.reset
+          pointer, matching_pointers = @itr_pointer.next(@builder['skip_matched_checkbutton'].active?, false)
+          if (pointer == nil)
+            notify('No more items to compare')
+          else
+            notify('Cycling items to compare from the end')
+          end
+        else
+          notify('')
+        end
+        if (pointer != nil)
+          load_comparison(pointer, matching_pointers)
+        end
+      end
+
+      # Load a new pointer and its matching pointers to be displayed
+      #
+      # Parameters::
+      # * *pointer* (<em>(_FileInfo_ | _SegmentPointer_)</em>): The pointer, or nil if none
+      # * *matching_pointers* (<em>list< [ (_FileInfo_ | _SegmentPointer_), MatchingIndexSinglePointer ] ></em>): The sorted list of matching pointers, along with their matching index information
+      def load_comparison(pointer, matching_pointers)
+        @idx_focused = nil
+        original_container = @builder['original_container']
+        matching_pointers_container = @builder['matching_container']
+        # Delete previous containers if needed
+        original_container.remove(original_container.children[0]) if (original_container.children.size > 0)
+        matching_pointers_container.each do |child_widget|
+          matching_pointers_container.remove(child_widget)
+        end
+        # Create new containers
+        original_container << @gui_controller.create_widget_for_matching_pointer(pointer, @matching_selection.matching_pointers[pointer] == pointer)
+        # For each encountered CRC, keep the matching pointer widget
+        # map< String, Gtk::Widget >
+        crcs = {}
+        matching_pointers.each do |matching_pointer, matching_file_info|
+          crc = matching_pointer.get_crc
+          if crcs.has_key?(crc)
+            # Add it to the existing widget
+            crcs[crc].add_pointer(matching_pointer)
+          else
+            matching_pointer_widget = @gui_controller.create_widget_for_matching_pointer(matching_pointer, @matching_selection.matching_pointers[pointer] == matching_pointer)
+            matching_pointers_container << matching_pointer_widget
+            crcs[crc] = matching_pointer_widget
+          end
+        end
+        # First, focus the original one
+        set_focused(-1)
+      end
 
       # Set the focused matching pointer
       #
@@ -89,6 +162,17 @@ module FilesRebuilder
           # Make sure it is visible
           @builder['matching_scrolledwindow'].vadjustment.value = matching_pointers_container.children[idx_focus].allocation.y if (idx_focus != -1)
         end
+      end
+
+      private
+
+      # Notify a message to the user
+      #
+      # Parameters::
+      # * *message* (_String_): Message to notify
+      def notify(message)
+        status_bar = @builder['statusbar']
+        status_bar.push(status_bar.get_context_id(''), message)
       end
 
     end
